apply: allow using maps,slices, and structs as arguments
Maps in particular can be thought of as functions from their keys to
their values. This allows them to be called like a function using
Apply. Likewise, slices can be thought of as functions from their
indicies to their values. Finally structs can be thought of as
functions from either public members or indicies to values.