dyn: Rework multiple function returns to be a tuple type
For multiple function returns one needs to be able to differentiate
between a function returning []interface{} and a function with
multiple returns. Compose is modified to use this new return type to
allow coposition of function returning multiple args and a a function
that takes the same number of args.